Organize packages and types
