+++ app/src/main/java/io/github/pietrocaselani/moviestraker/binders/MoviesListBindAdapter.kt
@@ -2,8 +2,8 @@ package io.github.pietrocaselani.moviestraker.binders
 
 import android.databinding.BindingAdapter
 import android.support.v7.widget.RecyclerView
-import io.github.pietrocaselani.moviestraker.upcoming.MovieListViewModel
-import io.github.pietrocaselani.moviestraker.upcoming.MoviesAdapter
+import io.github.pietrocaselani.moviestraker.ui.upcoming.MovieListViewModel
+import io.github.pietrocaselani.moviestraker.ui.upcoming.MoviesAdapter
 import jp.wasabeef.recyclerview.animators.SlideInDownAnimator
 
 /**
