+++ app/src/main/java/io/github/pietrocaselani/moviestraker/ui/MainActivity.kt
@@ -1,4 +1,4 @@
-package io.github.pietrocaselani.moviestraker
+package io.github.pietrocaselani.moviestraker.ui
 
 import android.os.Bundle
 import android.support.v4.app.Fragment
@@ -7,6 +7,7 @@ import android.support.v7.widget.Toolbar
 import dagger.android.AndroidInjector
 import dagger.android.DispatchingAndroidInjector
 import dagger.android.support.HasSupportFragmentInjector
+import io.github.pietrocaselani.moviestraker.R
 import javax.inject.Inject
 
 class MainActivity : AppCompatActivity(), HasSupportFragmentInjector {
