+++ app/src/main/java/io/github/pietrocaselani/moviestraker/ui/upcoming/EndlessRecyclerViewScrollListener.kt
@@ -1,4 +1,4 @@
-package io.github.pietrocaselani.moviestraker.upcoming
+package io.github.pietrocaselani.moviestraker.ui.upcoming
 
 import android.support.v7.widget.LinearLayoutManager
 import android.support.v7.widget.RecyclerView
