+++ app/src/main/java/io/github/pietrocaselani/moviestraker/ui/upcoming/MoviesAdapter.kt
@@ -1,4 +1,4 @@
-package io.github.pietrocaselani.moviestraker.upcoming
+package io.github.pietrocaselani.moviestraker.ui.upcoming
 
 import android.support.v7.widget.RecyclerView
 import android.view.LayoutInflater
