+++ app/src/main/java/io/github/pietrocaselani/moviestraker/ui/upcoming/UpcomingFragmentModule.kt
@@ -1,4 +1,4 @@
-package io.github.pietrocaselani.moviestraker.upcoming
+package io.github.pietrocaselani.moviestraker.ui.upcoming
 
 import dagger.Module
 import dagger.android.ContributesAndroidInjector
