+++ app/src/main/java/io/github/pietrocaselani/moviestraker/ui/upcoming/UpcomingViewModel.kt
@@ -1,4 +1,4 @@
-package io.github.pietrocaselani.moviestraker.upcoming
+package io.github.pietrocaselani.moviestraker.ui.upcoming
 
 import android.databinding.BaseObservable
 import android.databinding.ObservableField
