+++ app/src/main/res/layout/movie_list_item.xml
@@ -2,22 +2,22 @@
 <layout xmlns:android="http://schemas.android.com/apk/res/android"
         xmlns:app="http://schemas.android.com/apk/res-auto"
         xmlns:tools="http://schemas.android.com/tools"
-        tools:context="io.github.pietrocaselani.moviestraker.MainActivity">
+        tools:context="io.github.pietrocaselani.moviestraker.ui.MainActivity">
 
     <data>
 
         <variable
             name="viewModel"
-            type="io.github.pietrocaselani.moviestraker.upcoming.MovieListViewModel"/>
+            type="io.github.pietrocaselani.moviestraker.ui.upcoming.MovieListViewModel"/>
     </data>
 
     <android.support.v7.widget.CardView
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        app:cardUseCompatPadding="true"
-        app:cardElevation="@dimen/min_margin"
         app:cardBackgroundColor="@color/cardview_dark_background"
-        app:cardPreventCornerOverlap="false">
+        app:cardElevation="@dimen/min_margin"
+        app:cardPreventCornerOverlap="false"
+        app:cardUseCompatPadding="true">
 
         <LinearLayout
             android:layout_width="match_parent"
